approx second derivative
