+++ src/optimize/util.rs
@@ -1,6 +1,25 @@
-use num::complex::{Complex32, Complex64};
+use num::complex::{Complex32, Complex64, ComplexFloat};
 
 use crate::optimize::*;
+
+pub fn approx_derivative<F, C, M>(fun: F) -> impl Fn(C) -> C
+where
+    F: Fn(C) -> C,
+    C: IntoMetric<M> + ComplexFloat + Espilon,
+    M: Metric,
+{
+    let delta: C = C::epsilon().powf(C::from(1.0 / 3.0).unwrap().re());
+    move |x: C| (fun(x + delta) - fun(x)) / delta
+}
+pub fn approx_second_derivative<F, C, M>(fun: F) -> impl Fn(C) -> C
+where
+    F: Fn(C) -> C,
+    C: IntoMetric<M> + ComplexFloat + Espilon,
+    M: Metric,
+{
+    let delta: C = C::epsilon().powf(C::from(1.0 / 3.0).unwrap().re());
+    move |x: C| (fun(x + delta) - fun(x) * C::from(2.0).unwrap() + fun(x - delta)) / delta / delta
+}
 
 /// A trait implemented for `f32`,`f64`,`Complex32` and `Complex64`
 ///
