+++ src/special/kv.rs
@@ -16,7 +16,6 @@ pub fn kv(mut v: f64, mut z: Complex<f64>) -> Complex<f64> {
     if v.is_nan() {
         v = 0.0;
     }
-
     let res = complex_bessel_rs::bessel_k::bessel_k(v, z);
     if res.is_err() {
         println!("{v} {z}");
