+++ src/special/mod.rs
@@ -8,6 +8,6 @@ pub use trig::*;
 pub fn i0(x: Array1<f64>) -> Result<Array1<Complex<f64>>, i32> {
     // TODO implement this using cephes
     x.mapv(|v| complex_bessel_rs::bessel_i::bessel_i(0.0, v.into()))
-    .into_iter()
-    .collect()
+        .into_iter()
+        .collect()
 }
